@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "nx": "nx",
     "start": "ng serve",
     "build": "ng build",
@@ -31,58 +31,59 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.11",
-    "@angular/cdk": "13.3.9",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/forms": "13.3.11",
-    "@angular/material": "13.3.9",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
-    "@ngrx/effects": "13.0.2",
-    "@ngrx/entity": "13.0.2",
-    "@ngrx/router-store": "13.0.2",
-    "@ngrx/store": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
-    "@nrwl/angular": "13.10.6",
+    "@angular/animations": "14.2.6",
+    "@angular/cdk": "14.2.5",
+    "@angular/common": "14.2.6",
+    "@angular/compiler": "14.2.6",
+    "@angular/core": "14.2.6",
+    "@angular/forms": "14.2.6",
+    "@angular/material": "14.2.5",
+    "@angular/platform-browser": "14.2.6",
+    "@angular/platform-browser-dynamic": "14.2.6",
+    "@angular/router": "14.2.6",
+    "@ngrx/effects": "14.0.2",
+    "@ngrx/entity": "14.0.2",
+    "@ngrx/router-store": "14.0.2",
+    "@ngrx/store": "14.0.2",
+    "@ngrx/store-devtools": "14.0.2",
+    "@nrwl/angular": "14.8.6",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "zone.js": "0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.9",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.3.9",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
-    "@nrwl/cli": "13.10.6",
-    "@nrwl/cypress": "13.10.6",
-    "@nrwl/eslint-plugin-nx": "13.10.6",
-    "@nrwl/jest": "13.10.6",
-    "@nrwl/linter": "13.10.6",
-    "@nrwl/workspace": "13.10.6",
+    "@angular-devkit/build-angular": "14.2.6",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "~14.2.0",
+    "@angular/compiler-cli": "14.2.6",
+    "@angular/language-service": "14.2.6",
+    "@nrwl/cli": "14.8.6",
+    "@nrwl/cypress": "14.8.6",
+    "@nrwl/eslint-plugin-nx": "14.8.6",
+    "@nrwl/jest": "14.8.6",
+    "@nrwl/linter": "14.8.6",
+    "@nrwl/workspace": "14.8.6",
     "@types/faker": "^5.5.8",
-    "@types/jest": "27.0.2",
+    "@types/jest": "28.1.8",
     "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "5.18.0",
-    "@typescript-eslint/parser": "5.18.0",
+    "@typescript-eslint/eslint-plugin": "5.36.1",
+    "@typescript-eslint/parser": "5.36.1",
     "cypress": "^8.3.0",
     "dotenv": "10.0.0",
-    "eslint": "8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-cypress": "^2.11.3",
     "faker": "^5.5.3",
     "jasmine-marbles": "~0.8.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.1.1",
-    "prettier": "2.5.1",
-    "ts-jest": "27.0.5",
-    "ts-node": "~10.2.1",
-    "typescript": "4.6.4",
-    "nx": "13.10.6"
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.2",
+    "nx": "14.8.6",
+    "prettier": "2.6.2",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4"
   }
 }
